chore: fix warning

+++ src/permutation.hpp
@@ -121,7 +121,7 @@ public:
     /// @brief 輸出
     void print()
     {
-        for (int i = 0; i < records.size(); i++)
+        for (int i = 0; i < static_cast<int>(records.size()); i++)
         {
             cout << records[i] << endl;
         }
